+++ src/test/interpreters/activation/service.unit.test.ts
@@ -18,7 +18,7 @@ import { ProcessServiceFactory } from '../../../client/common/process/processFac
 import { IProcessService, IProcessServiceFactory } from '../../../client/common/process/types';
 import { TerminalHelper } from '../../../client/common/terminal/helper';
 import { ITerminalHelper } from '../../../client/common/terminal/types';
-import { ICurrentProcess } from '../../../client/common/types';
+import { ICurrentProcess, Resource } from '../../../client/common/types';
 import { getNamesAndValues } from '../../../client/common/utils/enum';
 import { Architecture, OSType } from '../../../client/common/utils/platform';
 import { EnvironmentVariablesProvider } from '../../../client/common/variables/environmentVariablesProvider';
@@ -48,7 +48,7 @@ suite('Interpreters Activation - Python Environment Variables', () => {
     let workspace: IWorkspaceService;
     let interpreterService: IInterpreterService;
     let onDidChangeEnvVariables: EventEmitter<Uri | undefined>;
-    let onDidChangeInterpreter: EventEmitter<void>;
+    let onDidChangeInterpreter: EventEmitter<Resource>;
     const pythonInterpreter: PythonEnvironment = {
         path: '/foo/bar/python.exe',
         version: new SemVer('3.6.6-final'),
@@ -68,7 +68,7 @@ suite('Interpreters Activation - Python Environment Variables', () => {
         interpreterService = mock(InterpreterService);
         workspace = mock(WorkspaceService);
         onDidChangeEnvVariables = new EventEmitter<Uri | undefined>();
-        onDidChangeInterpreter = new EventEmitter<void>();
+        onDidChangeInterpreter = new EventEmitter<Resource>();
         when(envVarsService.onDidEnvironmentVariablesChange).thenReturn(onDidChangeEnvVariables.event);
         when(interpreterService.onDidChangeInterpreter).thenReturn(onDidChangeInterpreter.event);
         when(interpreterService.getActiveInterpreter(anything())).thenResolve(interpreter);
